// packages/api-studio-features/src/generated/OAuth2Config.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GrantType } from "./GrantType";

export type OAuth2Config = {
  grantType: GrantType;
  authorizationUrl: string;
  tokenUrl: string;
  clientId: string;
  clientSecret: string;
  scopes: string;
};